requirements.txt
# Web scraping and API requests
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Data processing
pandas==2.2.3
python-dotenv==1.0.1

# AI and embeddings
openai==1.59.8

# Vector database
chromadb==0.5.23

# Progress tracking
tqdm==4.67.1

# Data validation
pydantic==2.10.6
